Split metric collection into other files
